+++ transform/src/parser/nodes/name/mod.rs
@@ -1,0 +1,16 @@
+// libs
+use swc_common::Span;
+use swc_ecma_ast::*;
+
+// helpers
+use crate::parser::utils::get_key_value_node;
+
+pub fn create_name(name: String, span: Span) -> Expr {
+    let kind = get_key_value_node("kind".into(), "Name".into());
+    let value = get_key_value_node("value".into(), name.into());
+    let name = ObjectLit {
+        span,
+        props: vec![kind, value],
+    };
+    Expr::Object(name)
+}
